Add files via upload
These files make up the Source folder (SRC). These are just for a simple TeleOp and a base Autonomous file that doesn't use sensors, just a timer. It has a camera and a gyroscope, but I didn't really use those for anything. This was made in 2022 WPILIBJ, so there will probably need to be changes in the code if this is used in future WPILIBJ  versions. There are comments throughout the files that try to explain how the code runs. If this is your first time with FRC coding, even if you have coded before, read the comments and code like a book. It will make creating your own code so much easier! Good luck!!!